@@ -53,40 +54,48 @@
   "description": "Standard library",
   "devDependencies": {
     "LiveScript": "1.3.x",
-    "eslint": "1.9.x",
-    "grunt": "0.4.x",
-    "grunt-cli": "0.1.x",
-    "grunt-contrib-clean": "0.6.x",
-    "grunt-contrib-copy": "0.8.x",
-    "grunt-contrib-uglify": "0.10.x",
-    "grunt-contrib-watch": "0.6.x",
-    "grunt-karma": "0.12.x",
-    "grunt-livescript": "0.5.x",
-    "karma": "0.13.x",
-    "karma-chrome-launcher": "0.2.x",
-    "karma-firefox-launcher": "0.1.x",
-    "karma-ie-launcher": "0.2.x",
-    "karma-phantomjs-launcher": "0.2.x",
-    "karma-qunit": "0.1.x",
-    "phantomjs": "1.9.x",
+    "es-observable-tests": "0.2.x",
+    "eslint": "3.1.x",
+    "grunt": "1.0.x",
+    "grunt-cli": "1.2.x",
+    "grunt-contrib-clean": "1.0.x",
+    "grunt-contrib-copy": "1.0.x",
+    "grunt-contrib-uglify": "1.0.x",
+    "grunt-contrib-watch": "1.0.x",
+    "grunt-karma": "2.0.x",
+    "grunt-livescript": "0.6.x",
+    "karma": "1.1.x",
+    "karma-chrome-launcher": "1.0.x",
+    "karma-firefox-launcher": "1.0.x",
+    "karma-ie-launcher": "1.0.x",
+    "karma-phantomjs-launcher": "1.0.x",
+    "karma-qunit": "1.1.x",
+    "phantomjs-prebuilt": "2.1.x",
     "promises-aplus-tests": "2.1.x",
-    "qunitjs": "1.23.x",
-    "webpack": "1.12.x"
+    "qunitjs": "2.0.x",
+    "temp": "0.8.x",
+    "webpack": "1.13.x"
   },
   "directories": {},
   "dist": {
-    "shasum": "652294c14651db28fa93bd2d5ff2983a4f08c636",
-    "tarball": "https://registry.npmjs.org/core-js/-/core-js-1.2.7.tgz"
+    "shasum": "4de911e667b0eae9124e34254b53aea6fc618d3e",
+    "tarball": "https://registry.npmjs.org/core-js/-/core-js-2.4.1.tgz"
   },
-  "gitHead": "11a0a09335b8afca5b489d3a503093d6f1fedd42",
+  "gitHead": "5e106f64c726edf2849f0babc9096ce6664b7368",
   "homepage": "https://github.com/zloirock/core-js#readme",
   "keywords": [
+    "ES3",
+    "ECMAScript 3",
     "ES5",
     "ECMAScript 5",
     "ES6",
+    "ES2015",
     "ECMAScript 6",
+    "ECMAScript 2015",
     "ES7",
+    "ES2016",
     "ECMAScript 7",
+    "ECMAScript 2016",
     "Harmony",
     "Strawman",
     "Map",
