@@ -56,60 +56,54 @@
     "url": "https://github.com/form-data/form-data/issues"
   },
   "dependencies": {
-    "async": "^1.5.2",
+    "asynckit": "^0.4.0",
     "combined-stream": "^1.0.5",
-    "mime-types": "^2.1.10"
+    "mime-types": "^2.1.11"
   },
   "description": "A library to create readable \"multipart/form-data\" streams. Can be used to submit forms and file uploads to other web applications.",
   "devDependencies": {
-    "codacy-coverage": "^1.1.3",
-    "coveralls": "^2.11.8",
-    "cross-spawn": "^2.1.5",
-    "eslint": "^2.4.0",
+    "coveralls": "^2.11.13",
+    "cross-spawn": "^4.0.0",
+    "eslint": "^3.5.0",
     "fake": "^0.2.2",
     "far": "^0.0.7",
     "formidable": "^1.0.17",
-    "istanbul": "^0.4.2",
-    "pre-commit": "^1.1.2",
-    "request": "^2.69.0",
-    "rimraf": "^2.5.2"
+    "in-publish": "^2.0.0",
+    "is-node-modern": "^1.0.0",
+    "istanbul": "^0.4.5",
+    "pkgfiles": "^2.3.0",
+    "pre-commit": "^1.1.3",
+    "request": "^2.74.0",
+    "rimraf": "^2.5.4"
   },
   "directories": {},
   "dist": {
-    "shasum": "05ac6bc22227b43e4461f488161554699d4f8b5e",
-    "tarball": "https://registry.npmjs.org/form-data/-/form-data-1.0.0-rc4.tgz"
+    "shasum": "6f0aebadcc5da16c13e1ecc11137d85f9b883b25",
+    "tarball": "https://registry.npmjs.org/form-data/-/form-data-2.0.0.tgz"
   },
   "engines": {
-    "node": ">= 0.10"
+    "node": ">= 0.12"
   },
-  "gitHead": "f73996e0508ee2d4b2b376276adfac1de4188ac2",
+  "gitHead": "652b16ff5b9077bdf65eb66b67286c823c2a1040",
   "homepage": "https://github.com/form-data/form-data#readme",
   "license": "MIT",
   "main": "./lib/form_data",
   "maintainers": [
     {
-      "name": "felixge",
-      "email": "felix@debuggable.com"
-    },
-    {
-      "name": "idralyuk",
-      "email": "igor@buran.us"
-    },
-    {
       "name": "alexindigo",
       "email": "iam@alexindigo.com"
     },
     {
-      "name": "mikeal",
-      "email": "mikeal.rogers@gmail.com"
-    },
-    {
-      "name": "celer",
-      "email": "dtyree77@gmail.com"
-    },
-    {
       "name": "dylanpiercey",
       "email": "pierceydylan@gmail.com"
+    },
+    {
+      "name": "felixge",
+      "email": "felix@debuggable.com"
+    },
+    {
+      "name": "mikeal",
+      "email": "mikeal.rogers@gmail.com"
     }
   ],
   "name": "form-data",
