@@ -1,29 +1,16 @@
-import * as types from './constants/actionTypes'
-import * as Actions from './actions/index'
 import config from 'config'
+import Parse from 'parse/node'
 import request from 'request'
 import HashMap  from 'hashmap'
+import objectAssign from 'object-assign'
 import FB from 'fb'
-import https from 'https'
-
-import * as _ from 'underscore'
-//import * as redux from 'redux'
-
-import Parse from 'parse/node'
-import {
-    app,
-    rules,
-    callSendAPI,
-    sendTypingOn,
-    sendTypingOff
-} from './bot'
-
-//GetProductsParams = require('./models/GetProductsParams')
-
-var ParseModels = require('./ParseModels');
-var Consumer = require('./ParseModels').Consumer;
-var User = require('./ParseModels').User;
-var ConsumerAddress = require('./ParseModels').ConsumerAddress;
+import thunk from 'redux-thunk';
+import * as bot from './bot'
+import * as types from './constants/actionTypes'
+import * as Actions from './actions/index'
+import * as ParseModels from './ParseModels';
+import { extractParseAttributes } from './ParseUtils'
+import { createStore, applyMiddleware } from 'redux';
 
 const APP_SECRET = (process.env.MESSENGER_APP_SECRET) ? process.env.MESSENGER_APP_SECRET : config.get('APP_SECRET');
 
@@ -50,10 +37,9 @@ FB.options({
     redirectUri:    REDIRECT_URI
 });
 
-console.log(FB)
+//console.log('Exist FB.init: '+( typeof FB.init === 'function') );
 
-global.Parse = Parse;
-global.FB = FB;
+
 
 //console.log(FB.getLoginUrl({ scope: 'user_about_me' }));
 
@@ -66,24 +52,122 @@ global.FB = FB;
  });
 */
 
-rules.set('hola', sendMenuMessage);
-rules.set('cuenta', sendBillMessage);
+bot.rules.set('hola', sendMenuMessage);
+bot.rules.set('cuenta', sendBillMessage);
 
 var order = new HashMap();
 
+const initialState = {
+    geocodedLocation: {lat: -1, lng: -1},
+    addresses: [],
+    mapBounds: {},
+    mapAddress: '',
+    useSubCategories: false,
+    products: [],
+    pointOfSales: [],
+    categories: [],
+    paymentMethods: [],
+    consumer: {},
+    orders: { ongoing: [], delivered: [] },
+    orderToRate: null,
+    currentOrder: {
+        consumerAddress:{},
+        items: []
+    },
+    locationZoom: 3,
+    profileIsOpen: false,
+    mapAddressIsOpen: false,
+    addressFormIsOpen: false,
+    addressListIsOpen: false,
+    savingAddress: false,
+    creatingOrder: false,
+    paymentMethodNotSelected: false,
+    cartTotalIsBelowMinimumPrice: false,
+    currentCategory: {},
+    consumerNotFound: false,
+    updatingConsumer: false,
+    currentUser: {},
+    pendingOrder: false
+};
+
+const reducer = (state = initialState, action) => {
+    console.log('ACTION');
+    console.log(action);
+    let item;
+    switch (action.type) {
+        case types.APP_LOADED:
+            console.log('Application loaded');
+            return {...state};
+        case types.CONSUMER_UPDATED:
+            return {...state};
+        case types.CONSUMER_LOADED:
+            if (state.consumerNotFound) delete state.consumerNotFound
+            let consumer = extractParseAttributes(action.data.consumer)
+            //objectAssign(state.consumerAddress, {consumer})
+            objectAssign(state, {consumer, currentUser: action.data.user});
+            return {...state};
+        case types.CONSUMER_NOT_FOUND:
+            objectAssign(state, {
+                consumerNotFound: true,
+                currentUser: action.data.user
+            })
+            return {...state};
+        case types.CONSUMER_ADDRESSES_LOADED:
+            let addresses = action.data.map(a => extractParseAttributes(a))
+            objectAssign(state, {addresses: addresses})
+            return {...state};
+        case types.RENDER_MENU:
+            console.log('RENDER_MENU');
+            var consumer = store.getState().consumer;
+            console.log('Store');
+            console.log(store.getState())
+            console.log('Consumer');
+            console.log(consumer);
+
+            request({
+                uri: 'https://graph.facebook.com/v2.6/' + '1100195690052041',
+                qs: {access_token: PAGE_ACCESS_TOKEN, fields: 'first_name,last_name,locale,timezone,gender'},
+                method: 'GET'
+            }, renderMenuMessage);
+
+            return state;
+        case types.CONSUMER_ORDERS_LOADED:
+            console.log('CONSUMER_ORDERS_LOADED');
+            const orders = action.data
+            let ongoing = orders['ongoing'].map(o => extractParseAttributes(o))
+            let delivered = orders['delivered'].map(o => extractParseAttributes(o))
+            let orderToRate = delivered[0]
+            objectAssign(state, {
+                orders: {ongoing, delivered},
+                orderToRate
+            })
+            return {...state}
+        default:
+            console.log(action);
+            return state;
+    }
+    console.log('\n');
+}
+
+var store = createStore(reducer, applyMiddleware(thunk));
+
+store.subscribe(() =>
+    console.log('\n') //store.getState())
+);
+
+global.Parse = Parse;
+global.FB = FB;
+global.store = store;
+
 function sendMenuMessage(recipientId) {
     console.log('Typing ON');
-    sendTypingOn(recipientId);
+    bot.sendTypingOn(recipientId);
 
-    new Parse.Query(User).get('DH6JIZFrGW').then(
+    new Parse.Query(ParseModels.User).get('DH6JIZFrGW').then(
         function (user) {
-            var login = Actions.loadConsumer(user, {});
-            console.log('user');
-            console.log(user);
-            console.log('login');
-            console.log(login());
+            store.dispatch(Actions.loadConsumer(user, store.dispatch));
 
-            new Parse.Query(Consumer).equalTo('user', user).first().then(consumer => {
+            /*new Parse.Query(ParseModels.Consumer).equalTo('user', user).first().then(consumer => {
                 if (consumer) {
                     request({
                         uri: 'https://graph.facebook.com/v2.6/' + recipientId,
@@ -93,7 +177,7 @@ function sendMenuMessage(recipientId) {
                 }
             }).fail(error => {
                 console.log(error);
-            })
+            })*/
         },
         function (object, error) {
             console.log(error);
@@ -103,7 +187,7 @@ function sendMenuMessage(recipientId) {
 
 function listCategories(recipientId, catIdx){
     console.log('Typing ON');
-    sendTypingOn(recipientId);
+    bot.sendTypingOn(recipientId);
 
     Parse.Cloud.run('getProducts', { businessId: BUSINESS_ID }).then(function(result){
             var elements = splitCategories(result.categories, catIdx);
@@ -145,8 +229,8 @@ function listCategories(recipientId, catIdx){
             };
 
             console.log('Typing OFF');
-            sendTypingOff(recipientId);
-            callSendAPI(messageData);
+            bot.sendTypingOff(recipientId);
+            bot.callSendAPI(messageData);
 
         },
         function(error) {
@@ -155,7 +239,7 @@ function listCategories(recipientId, catIdx){
 }
 
 function listProducts(recipientId, category, proIdx){
-    sendTypingOn(recipientId);
+    bot.sendTypingOn(recipientId);
 
     Parse.Cloud.run('getProducts', { businessId: BUSINESS_ID, category: category }).then(function(result) {
 
@@ -192,8 +276,8 @@ function listProducts(recipientId, category, proIdx){
                 }
             };
 
-            sendTypingOff(recipientId);
-            callSendAPI(messageData);
+            bot.sendTypingOff(recipientId);
+            bot.callSendAPI(messageData);
         },
         function(error) {
 
@@ -214,7 +298,7 @@ function addProduct(id){
 
 function sendBillMessage(recipientId){
     // Generate a random receipt ID as the API requires a unique ID
-    sendTypingOn(recipientId);
+    bot.sendTypingOn(recipientId);
 
     var receiptId = "order" + Math.floor(Math.random()*1000);
     var elements = [];
@@ -305,8 +389,8 @@ function sendBillMessage(recipientId){
                             };
                             order = new HashMap();
                             console.log("callSendAPI(messageData)");
-                            sendTypingOff(recipientId);
-                            callSendAPI(messageData);
+                            bot.sendTypingOff(recipientId);
+                            bot.callSendAPI(messageData);
 
                         }
                     });
@@ -438,8 +522,8 @@ function renderMenuMessage(error, response, body) {
                     }
                 };
 
-                sendTypingOff(recipientId);
-                callSendAPI(messageData);
+                bot.sendTypingOff(recipientId);
+                bot.callSendAPI(messageData);
             },
             function(error) {
                 console.log("Lookup failed");
@@ -449,12 +533,4 @@ function renderMenuMessage(error, response, body) {
     }
 }
 
-/**
- * CONSUMER ADDRESS LOADED action
- */
-export function consumerAddressesLoaded(consumerAddresses) {
-    return {
-        type: types.CONSUMER_ADDRESSES_LOADED,
-        data: consumerAddresses
-    }
-}
+store.dispatch({type: types.APP_LOADED})
