@@ -1,6 +1,7 @@
-// ParseModels.js
+import Parse from 'parse'
+import ConsumerAddress from './models/ConsumerAddress'
 
-const Parse = require('parse/node');
+Parse.Object.registerSubclass('ConsumerAddress', ConsumerAddress);
 
 const User = Parse.Object.extend('User');
 const Customer = Parse.Object.extend('Customer');
@@ -15,7 +16,6 @@ const OrderItem = Parse.Object.extend('OrderItem');
 const OrderItemModifier = Parse.Object.extend('OrderItemModifier');
 const OrderItemModifierGroup = Parse.Object.extend('OrderItemModifierGroup');
 const PaymentMethod = Parse.Object.extend('PaymentMethod');
-const ConsumerAddress = Parse.Object.extend('ConsumerAddress');
 
 module.exports = {
     User,
