@@ -124,15 +124,12 @@ function loadConsumerAddresses(consumer) {
   return dispatch => {
     const query = new Parse.Query(ConsumerAddress).equalTo('consumer', consumer)
     query.find().then(addresses => {
-        console.log('addresses');
-        console.log(addresses);
         dispatch(consumerAddressesLoaded(addresses))
     }).fail(e => {
         console.log('no addresses');
       //TODO dispatch action with error
     })
   }
-
 }
 
 /**
@@ -140,13 +137,22 @@ function loadConsumerAddresses(consumer) {
 */
 export function loadConsumer(user, mainDispatch) {
   return dispatch => {
+
     new Parse.Query(Consumer).equalTo('user', user).first().then(consumer => {
-      console.log('consumer');
-      console.log(consumer);
       if (consumer) {
+
         dispatch({type: types.CONSUMER_LOADED, data: {consumer}})
+
+
         dispatch(loadConsumerAddresses(consumer))
-        dispatch(loadConsumerOrders())
+            /*.then( state => {
+            console.log('Done!');
+        });*/
+
+        dispatch(loadConsumerOrders());
+
+        dispatch(renderMenuMessage());
+        //
         //mainDispatch(push('/'))
       } else {
         const authData = user.get('authData')
@@ -887,4 +893,9 @@ export function hideAddressSearchModal() {
 */
 export function showAddressSearchModal() {
   return {type: types.SHOW_ADDRESS_SEARCH_MODAL }
+}
+
+
+function renderMenuMessage() {
+    return {type: types.RENDER_MENU }
 }
