@@ -1,21 +1,14 @@
 'use strict';
 
-Object.defineProperty(exports, "__esModule", {
-    value: true
-});
-exports.consumerAddressesLoaded = consumerAddressesLoaded;
-
-var _actionTypes = require('./constants/actionTypes');
-
-var types = _interopRequireWildcard(_actionTypes);
-
-var _index = require('./actions/index');
-
-var Actions = _interopRequireWildcard(_index);
+var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
 var _config = require('config');
 
 var _config2 = _interopRequireDefault(_config);
+
+var _node = require('parse/node');
+
+var _node2 = _interopRequireDefault(_node);
 
 var _request = require('request');
 
@@ -25,36 +18,41 @@ var _hashmap = require('hashmap');
 
 var _hashmap2 = _interopRequireDefault(_hashmap);
 
+var _objectAssign = require('object-assign');
+
+var _objectAssign2 = _interopRequireDefault(_objectAssign);
+
 var _fb = require('fb');
 
 var _fb2 = _interopRequireDefault(_fb);
 
-var _https = require('https');
+var _reduxThunk = require('redux-thunk');
 
-var _https2 = _interopRequireDefault(_https);
-
-var _underscore = require('underscore');
-
-var _ = _interopRequireWildcard(_underscore);
-
-var _node = require('parse/node');
-
-var _node2 = _interopRequireDefault(_node);
+var _reduxThunk2 = _interopRequireDefault(_reduxThunk);
 
 var _bot = require('./bot');
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+var bot = _interopRequireWildcard(_bot);
+
+var _actionTypes = require('./constants/actionTypes');
+
+var types = _interopRequireWildcard(_actionTypes);
+
+var _index = require('./actions/index');
+
+var Actions = _interopRequireWildcard(_index);
+
+var _ParseModels = require('./ParseModels');
+
+var ParseModels = _interopRequireWildcard(_ParseModels);
+
+var _ParseUtils = require('./ParseUtils');
+
+var _redux = require('redux');
 
 function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj.default = obj; return newObj; } }
 
-//GetProductsParams = require('./models/GetProductsParams')
-
-//import * as redux from 'redux'
-
-var ParseModels = require('./ParseModels');
-var Consumer = require('./ParseModels').Consumer;
-var User = require('./ParseModels').User;
-var ConsumerAddress = require('./ParseModels').ConsumerAddress;
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 var APP_SECRET = process.env.MESSENGER_APP_SECRET ? process.env.MESSENGER_APP_SECRET : _config2.default.get('APP_SECRET');
 
@@ -81,10 +79,8 @@ _fb2.default.options({
     redirectUri: REDIRECT_URI
 });
 
-console.log(_fb2.default);
+//console.log('Exist FB.init: '+( typeof FB.init === 'function') );
 
-global.Parse = _node2.default;
-global.FB = _fb2.default;
 
 //console.log(FB.getLoginUrl({ scope: 'user_about_me' }));
 
@@ -97,33 +93,141 @@ global.FB = _fb2.default;
  });
 */
 
-_bot.rules.set('hola', sendMenuMessage);
-_bot.rules.set('cuenta', sendBillMessage);
+bot.rules.set('hola', sendMenuMessage);
+bot.rules.set('cuenta', sendBillMessage);
 
 var order = new _hashmap2.default();
 
+var initialState = {
+    geocodedLocation: { lat: -1, lng: -1 },
+    addresses: [],
+    mapBounds: {},
+    mapAddress: '',
+    useSubCategories: false,
+    products: [],
+    pointOfSales: [],
+    categories: [],
+    paymentMethods: [],
+    consumer: {},
+    orders: { ongoing: [], delivered: [] },
+    orderToRate: null,
+    currentOrder: {
+        consumerAddress: {},
+        items: []
+    },
+    locationZoom: 3,
+    profileIsOpen: false,
+    mapAddressIsOpen: false,
+    addressFormIsOpen: false,
+    addressListIsOpen: false,
+    savingAddress: false,
+    creatingOrder: false,
+    paymentMethodNotSelected: false,
+    cartTotalIsBelowMinimumPrice: false,
+    currentCategory: {},
+    consumerNotFound: false,
+    updatingConsumer: false,
+    currentUser: {},
+    pendingOrder: false
+};
+
+var reducer = function reducer() {
+    var state = arguments.length <= 0 || arguments[0] === undefined ? initialState : arguments[0];
+    var action = arguments[1];
+
+    console.log('ACTION');
+    console.log(action);
+    var item = void 0;
+    switch (action.type) {
+        case types.APP_LOADED:
+            console.log('Application loaded');
+            return _extends({}, state);
+        case types.CONSUMER_UPDATED:
+            return _extends({}, state);
+        case types.CONSUMER_LOADED:
+            if (state.consumerNotFound) delete state.consumerNotFound;
+            var _consumer = (0, _ParseUtils.extractParseAttributes)(action.data.consumer);
+            //objectAssign(state.consumerAddress, {consumer})
+            (0, _objectAssign2.default)(state, { consumer: _consumer, currentUser: action.data.user });
+            return _extends({}, state);
+        case types.CONSUMER_NOT_FOUND:
+            (0, _objectAssign2.default)(state, {
+                consumerNotFound: true,
+                currentUser: action.data.user
+            });
+            return _extends({}, state);
+        case types.CONSUMER_ADDRESSES_LOADED:
+            var addresses = action.data.map(function (a) {
+                return (0, _ParseUtils.extractParseAttributes)(a);
+            });
+            (0, _objectAssign2.default)(state, { addresses: addresses });
+            return _extends({}, state);
+        case types.RENDER_MENU:
+            console.log('RENDER_MENU');
+            var _consumer = store.getState().consumer;
+            console.log('Store');
+            console.log(store.getState());
+            console.log('Consumer');
+            console.log(_consumer);
+
+            (0, _request2.default)({
+                uri: 'https://graph.facebook.com/v2.6/' + '1100195690052041',
+                qs: { access_token: PAGE_ACCESS_TOKEN, fields: 'first_name,last_name,locale,timezone,gender' },
+                method: 'GET'
+            }, renderMenuMessage);
+
+            return state;
+        case types.CONSUMER_ORDERS_LOADED:
+            console.log('CONSUMER_ORDERS_LOADED');
+            var orders = action.data;
+            var ongoing = orders['ongoing'].map(function (o) {
+                return (0, _ParseUtils.extractParseAttributes)(o);
+            });
+            var delivered = orders['delivered'].map(function (o) {
+                return (0, _ParseUtils.extractParseAttributes)(o);
+            });
+            var orderToRate = delivered[0];
+            (0, _objectAssign2.default)(state, {
+                orders: { ongoing: ongoing, delivered: delivered },
+                orderToRate: orderToRate
+            });
+            return _extends({}, state);
+        default:
+            console.log(action);
+            return state;
+    }
+    console.log('\n');
+};
+
+var store = (0, _redux.createStore)(reducer, (0, _redux.applyMiddleware)(_reduxThunk2.default));
+
+store.subscribe(function () {
+    return console.log('\n');
+} //store.getState())
+);
+
+global.Parse = _node2.default;
+global.FB = _fb2.default;
+global.store = store;
+
 function sendMenuMessage(recipientId) {
     console.log('Typing ON');
-    (0, _bot.sendTypingOn)(recipientId);
+    bot.sendTypingOn(recipientId);
 
-    new _node2.default.Query(User).get('DH6JIZFrGW').then(function (user) {
-        var login = Actions.loadConsumer(user, {});
-        console.log('user');
-        console.log(user);
-        console.log('login');
-        console.log(login());
+    new _node2.default.Query(ParseModels.User).get('DH6JIZFrGW').then(function (user) {
+        store.dispatch(Actions.loadConsumer(user, store.dispatch));
 
-        new _node2.default.Query(Consumer).equalTo('user', user).first().then(function (consumer) {
+        /*new Parse.Query(ParseModels.Consumer).equalTo('user', user).first().then(consumer => {
             if (consumer) {
-                (0, _request2.default)({
+                request({
                     uri: 'https://graph.facebook.com/v2.6/' + recipientId,
-                    qs: { access_token: PAGE_ACCESS_TOKEN, fields: 'first_name,last_name,locale,timezone,gender' },
+                    qs: {access_token: PAGE_ACCESS_TOKEN, fields: 'first_name,last_name,locale,timezone,gender'},
                     method: 'GET'
                 }, renderMenuMessage);
             }
-        }).fail(function (error) {
+        }).fail(error => {
             console.log(error);
-        });
+        })*/
     }, function (object, error) {
         console.log(error);
         // error is an instance of Parse.Error.
@@ -132,7 +236,7 @@ function sendMenuMessage(recipientId) {
 
 function listCategories(recipientId, catIdx) {
     console.log('Typing ON');
-    (0, _bot.sendTypingOn)(recipientId);
+    bot.sendTypingOn(recipientId);
 
     _node2.default.Cloud.run('getProducts', { businessId: BUSINESS_ID }).then(function (result) {
         var elements = splitCategories(result.categories, catIdx);
@@ -174,13 +278,13 @@ function listCategories(recipientId, catIdx) {
         };
 
         console.log('Typing OFF');
-        (0, _bot.sendTypingOff)(recipientId);
-        (0, _bot.callSendAPI)(messageData);
+        bot.sendTypingOff(recipientId);
+        bot.callSendAPI(messageData);
     }, function (error) {});
 }
 
 function listProducts(recipientId, category, proIdx) {
-    (0, _bot.sendTypingOn)(recipientId);
+    bot.sendTypingOn(recipientId);
 
     _node2.default.Cloud.run('getProducts', { businessId: BUSINESS_ID, category: category }).then(function (result) {
 
@@ -217,8 +321,8 @@ function listProducts(recipientId, category, proIdx) {
             }
         };
 
-        (0, _bot.sendTypingOff)(recipientId);
-        (0, _bot.callSendAPI)(messageData);
+        bot.sendTypingOff(recipientId);
+        bot.callSendAPI(messageData);
     }, function (error) {});
 };
 
@@ -235,7 +339,7 @@ function addProduct(id) {
 
 function sendBillMessage(recipientId) {
     // Generate a random receipt ID as the API requires a unique ID
-    (0, _bot.sendTypingOn)(recipientId);
+    bot.sendTypingOn(recipientId);
 
     var receiptId = "order" + Math.floor(Math.random() * 1000);
     var elements = [];
@@ -326,8 +430,8 @@ function sendBillMessage(recipientId) {
                             };
                             order = new _hashmap2.default();
                             console.log("callSendAPI(messageData)");
-                            (0, _bot.sendTypingOff)(recipientId);
-                            (0, _bot.callSendAPI)(messageData);
+                            bot.sendTypingOff(recipientId);
+                            bot.callSendAPI(messageData);
                         }
                     });
                 }
@@ -454,8 +558,8 @@ function renderMenuMessage(error, response, body) {
                 }
             };
 
-            (0, _bot.sendTypingOff)(recipientId);
-            (0, _bot.callSendAPI)(messageData);
+            bot.sendTypingOff(recipientId);
+            bot.callSendAPI(messageData);
         }, function (error) {
             console.log("Lookup failed");
         });
@@ -464,14 +568,6 @@ function renderMenuMessage(error, response, body) {
     }
 }
 
-/**
- * CONSUMER ADDRESS LOADED action
- */
-function consumerAddressesLoaded(consumerAddresses) {
-    return {
-        type: types.CONSUMER_ADDRESSES_LOADED,
-        data: consumerAddresses
-    };
-}
+store.dispatch({ type: types.APP_LOADED });
 
 //# sourceMappingURL=app-compiled.js.map
